+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add ID-based route for gallery.
 
+### Changed
+
+- Display handle with '@' in account link tooltips.
+
 ## [2.13.0] - 2024-03-12
 
 ### Added